初步完成歌词部分；

+++ app/src/main/java/ayaseruri/torr/torrfm/objectholder/LrcInfo.java
@@ -4,17 +4,17 @@ package ayaseruri.torr.torrfm.objectholder;
  * Created by ayaseruri on 15/12/13.
  */
 public class LrcInfo {
-    private int time;
+    private long time;
     private String jp;
     private String cn;
 
-    public LrcInfo(int time, String jp, String cn) {
+    public LrcInfo(long time, String jp, String cn) {
         this.time = time;
         this.jp = jp;
         this.cn = cn;
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
     }
